fix(web): engine adapter never imports app code, desktop bundle builds in PR CI
The desktop app bundles packages/web's engine adapter through a Vite alias
that cannot resolve `@houston/app/*`; only the web build can. Three recent
adapter files imported app classifiers that way (feed-output.ts,
cp/fetch.ts), so every Tauri build of cloud-v0.6.16 failed on
"Rollup failed to resolve import" while the web build and PR CI passed.

- Move the three dependency-free modules (engine-waking-error,
  network-transport-error, waking-stuck-tracker) into the adapter, beside
  the transport that parses the same answers; app/src/lib keeps thin
  re-exports so app call sites and node:test entry points are unchanged and
  the tracker stays one shared instance.
- PR CI's app job now runs the desktop bundle build (tsgo + vite build),
  the only build that catches this class of import.
